Move ConstituencyLockView to LockBallotView, update URL
A future commit will move this View to using ballot IDs rather than
POSTing Post IDs to an Election endpoint.
